Attempt to implement the viewer atop the original VexFlow
I attempted to reimplement @ringw's viewer (as seen in
http://ringwa.lt/vexflow-demo/) atop the original (@0xFE's) VexFlow. The
original plan was to copy as much of @ringw's MusicXML parser as possible to
implement parsing into an internal representation that could be easily edited
and passed to VexFlow for rendering. However, @ringw's version made too many
modifications to VexFlow to do this easily.